% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oXim-main.R
\name{createFilterSetting}
\alias{createFilterSetting}
\title{Create filter-settings Object.}
\usage{
createFilterSetting(name = NULL, type = NULL, radius = NULL,
  times = NULL, tolerance = NULL)
}
\arguments{
\item{name}{Parameter to indicate prefixed profile of settings. This parameter has priority over the others.}

\item{type}{Indicates type of filter to use. See details below.}

\item{radius}{Indicates the size (on pixels) of sides of square used to apply the filters.}

\item{times}{Indicates number of times to apply the filters.}

\item{tolerance}{For \code{.noiselessFilter}, this parameter indicates proportion of pixels to consider from
filter matrix (radius x radius).}
}
\description{
This function allows to create correctly a filter-settings object in order to insert as input on
\code{getOxyrange} function.
}
\examples{
# Use default profile
createFilterSetting(name = "default")

# Generate a personalized profile
createFilterSetting(type = ".definerFilter", radius = c(3, 5, 5))
}

