% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LXOM-main.R
\name{echogramPlot.all}
\alias{echogramPlot.all}
\title{Wrap of \code{echogramPlot}.}
\usage{
echogramPlot.all(echogramOutputs, col = NULL, ...)
}
\arguments{
\item{echogramOutput}{Object of class \code{oxyclineData} with internal echogram matrix to be plotted.}

\item{colEchogram}{Pallete of colours to plot the echograms. If \code{NULL} (default) the system
will use the same combination used on object \code{colPallete}.}
}
\description{
Plot all echograms on an oxyclineData-class object.
}
\examples{
echogramPlot.all(fluidMatrix)
}

