% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oXim-main.R
\name{echogramPlot}
\alias{echogramPlot}
\title{Plot a matrix of a filtered echogram.}
\usage{
echogramPlot(echogramOutput, colEchogram = NULL, ...)
}
\arguments{
\item{echogramOutput}{Object of class \code{oxyclineData} with internal echogram matrix to be plotted.}

\item{colEchogram}{Pallete of colours to plot the echograms. If \code{NULL} (default) the system
will use the same combination used on object \code{colPallete}.}
}
\description{
This function uses an oxyclineData-class object and plot .
}
\examples{
echogramPlot(fluidMatrix)
}

