% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LXOM-main.R
\name{getLine98}
\alias{getLine98}
\title{Takes a matrix of echogram and calculate Oxycline.}
\usage{
getLine98(fluidMatrix, combinations = NULL, stepBYstep = TRUE)
}
\arguments{
\item{fluidMatrix}{Matrix or list of matrix of echograms.}

\item{combinations}{List with combination of filters.}

\item{stepBYstep}{Returns each echogram on a list.}

\item{intoOriginal}{Make the indexation on the original matrix?}
}
\description{
This function search othe Oxycline limits using filters and the procedure
especified in paper of Lau-Medrano W. and Oliveros-Ramos R..
}
\details{
getLine98

If \code{combinations} parameter is \code{NULL}, LXOM will use as default
\code{data.frame(type = c(".noiselessFilter", ".definerFilter"), radius = c(3, 3), times = c(2, 1), tolerance = c(0.2, NA),  stringsAsFactors = FALSE)}
}
\examples{
getLine98(fluidMatrix)
}

