% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{plot.echoData}
\alias{plot.echoData}
\title{Plot method for echoData}
\usage{
\method{plot}{echoData}(x, ...)
}
\arguments{
\item{x}{\code{echoData} object provided by \code{readEchograms} function.}

\item{...}{Arguments passed from \code{plot} function.}
}
\description{
This function takes an \code{echoData} object and plots an interpolated map
showing oxycline depth along shore.
}
\examples{
fileMode <- list(fish38_file   = system.file("extdata", "fish38.mat", package = "oXim"),
                 fluid120_file = system.file("extdata", "fluid120.mat", package = "oXim"),
                 blue38_file   = system.file("extdata", "blue38.mat", package = "oXim"))
echoData <- readEchograms(fileMode = fileMode)
print(echoData)
}

