% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{plot.oxyclineData}
\alias{plot.oxyclineData}
\title{Plot method for oxyclineData}
\usage{
\method{plot}{oxyclineData}(x, interpParams = list(duplicate = "strip"),
  coastline = TRUE, mapParams = list(database = "world"), ...)
}
\arguments{
\item{x}{Object of class \code{oxyclineData}}

\item{interpParams}{\code{list} object including parameters passed to \code{\link{interp}} function.}

\item{coastline}{\code{logical}. Do you want to show the coast line over the map?}

\item{mapParams}{If \code{coastline=TRUE}, \code{list} object including parameters passed to \code{\link{map}}
function.}

\item{...}{Extra arguments passed to \code{\link{image}} function.}
}
\description{
This method takes an \code{oxyclineData} object, make an interpolation of oxycline
values and show them on a map.
}

