% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{print.echoData}
\alias{print.echoData}
\title{Print method for echoData Objects.}
\usage{
\method{print}{echoData}(x, ...)
}
\arguments{
\item{x}{\code{echoData} object provided by \code{readEchograms} function.}

\item{...}{Extra argumemts.}
}
\description{
Shows main information from echodata Objects.
}
\examples{
fileMode <- list(fish38_file   = system.file("extdata", "fish38.mat", package = "oXim"),
                 fluid120_file = system.file("extdata", "fluid120.mat", package = "oXim"),
                 blue38_file   = system.file("extdata", "blue38.mat", package = "oXim"))
echoData <- readEchograms(fileMode = fileMode)
print(echoData)
}

