% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{print.summary.echoData}
\alias{print.summary.echoData}
\title{Print method for summary.echoData}
\usage{
\method{print}{summary.echoData}(x)
}
\arguments{
\item{x}{\code{echoData.summary} object provided by application of summary method to \code{echoData} object.}
}
\description{
Shows main information from echodata.summary object.
}
\examples{
fileMode <- list(fish38_file   = system.file("extdata", "fish38.mat", package = "oXim"),
                 fluid120_file = system.file("extdata", "fluid120.mat", package = "oXim"),
                 blue38_file   = system.file("extdata", "blue38.mat", package = "oXim"))
echoData <- readEchograms(fileMode = fileMode)

mySummary <- summary(echoData)
print(mySummary)
}

