% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LXOM-main.R
\name{readEchograms}
\alias{readEchograms}
\title{Takes outputs from Echopen and generates a matrix to calculate Oxycline.}
\usage{
readEchograms(directory, validFish38 = c(-100, -21), validBlue38 = c(-100,
  -56), upLimitFluid120 = -53, pinInterval = 50,
  date.format = "\%d-\%m-\%Y \%H:\%M:\%S")
}
\arguments{
\item{directory}{Directory where are outputs of Echoopen for Fluid-like, Blue noise 
and Fish.}

\item{validFish38}{Range of valid values for Fish-38kHz}

\item{validBlue38}{Range of valid values for Blue-38kHz}

\item{upLimitFluid120}{Upper limit for Fluidlike-120kHz}

\item{pinInterval}{Time threshold (in secs) to consider separate two matrices (echograms).}

\item{date.format}{A character string. The default method is \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
}
\description{
This function search outputs of Echoopen for Fluid-like, Blue noise 
and Fish and use them to make a filtered matrix to calculate the Oxycline limits.
}
\details{
readEchograms
}
\examples{
readEchograms(directory = "C:/Echodata")
}

