% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{summary.echoData}
\alias{summary.echoData}
\title{Summary method for echoData}
\usage{
\method{summary}{echoData}(x)
}
\arguments{
\item{x}{\code{echoData} object provided by \code{readEchograms} function.}
}
\description{
Get summary information of echograms included on echodata objects.
}
\examples{
fileMode <- list(fish38_file   = system.file("extdata", "fish38.mat", package = "oXim"),
                 fluid120_file = system.file("extdata", "fluid120.mat", package = "oXim"),
                 blue38_file   = system.file("extdata", "blue38.mat", package = "oXim"))
echoData <- readEchograms(fileMode = fileMode)
mySummary <- summary(echoData)
}

