% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{summary.oxyclineData}
\alias{summary.oxyclineData}
\title{Summary method for oxyclineData}
\usage{
\method{summary}{oxyclineData}(x)
}
\arguments{
\item{x}{Object of class \code{oxyclineData}.}
}
\description{
Summary method for oxyclineData
}

