% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LXOM-main.R
\name{echogramPlot}
\alias{echogramPlot}
\title{Takes a list of filtered echograms plot them.}
\usage{
echogramPlot(echogramOutput, col = NULL, ...)
}
\arguments{
\item{echogramOutput}{Object of class \code{oxyclineData} with internal echogram matrix to be plotted.}

\item{col}{Pallete of colours to plot the echograms. If \code{NULL} (default) the system
will use the same combination used on IMARPE.}
}
\description{
This function uses an oxyclineData-class object and plot .
}
\examples{
echogramPlot(fluidMatrix)
}

