% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oXim-main.R
\name{getOxyrange}
\alias{getOxyrange}
\title{Takes a matrix of echogram and calculate Oxycline.}
\usage{
getOxyrange(fluidMatrix, filterSettings = NULL, stepBYstep = FALSE)
}
\arguments{
\item{fluidMatrix}{Matrix or list of matrix of echograms.}

\item{filterSettings}{List with combination of filters.}

\item{stepBYstep}{Returns each echogram on a list.}
}
\description{
This function search othe Oxycline limits using filters and the procedure
especified in paper of Lau-Medrano W. and Oliveros-Ramos R..
}
\details{
If \code{filterSettings} parameter is \code{NULL}, oXim will use as default \code{data.frame(type = c(".noiselessFilter", ".definerFilter"), radius = c(3, 3), times = c(2, 1), tolerance = c(0.2, NA), stringsAsFactors = FALSE)}
}
\examples{
fileMode <- list(fish38_file   = system.file("extdata", "fish38.mat", package = "oXim"),
                 fluid120_file = system.file("extdata", "fluid120.mat", package = "oXim"),
                 blue38_file   = system.file("extdata", "blue38.mat", package = "oXim"))
echoData <- readEchograms(fileMode = fileMode)
oxyclineRange <- getOxyrange(echoData)
}

