% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{plot.oxyclineData}
\alias{plot.oxyclineData}
\title{Plot method for oxyclineData}
\usage{
\method{plot}{oxyclineData}(x, what = seq_along(x$outputs),
  plot.oxyrange = TRUE, col = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{oxyclineData}}

\item{what}{What echogram (in number) do you want to plot?}

\item{plot.oxyrange}{\code{logical}. If \code{TRUE} (default), show lines of oxycline range.}

\item{col}{Vector of color (palette) used to plot echograms.}

\item{...}{Extra arguments passed to \code{echogramPlot} function.}
}
\description{
Plot method for oxyclineData
}
\examples{
# Set a list of directories
fileMode <- list(fish38_file   = system.file("extdata", "fish38.mat", package = "oXim"),
                 fluid120_file = system.file("extdata", "fluid120.mat", package = "oXim"),
                 blue38_file   = system.file("extdata", "blue38.mat", package = "oXim"))

# Read echograms
echoData <- readEchograms(fileMode = fileMode)

# Calculate oxycline range
echoOutputs <- getOxyrange(fluidMatrix = echoData)

# Plot oxycline range
plot(echoOutputs)
}

