% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{plot.oxyclineData}
\alias{plot.oxyclineData}
\title{Plot method for oxyclineData}
\usage{
\method{plot}{oxyclineData}(x, interpParams = list(myGrid = NULL), ...)
}
\arguments{
\item{x}{Object of class \code{oxyclineData}}

\item{interpParams}{\code{list} object including parameters passed to \code{\link{idw}} function.}

\item{...}{Extra arguments passed to \code{\link{plot.SpatialGridDataFrame}} function.}
}
\description{
This method takes an \code{oxyclineData} object, make an interpolation of oxycline
values and show them on a map. Interpolation methodology is based on \code{akima} package.
}

