% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{summary.oxyclineData}
\alias{summary.oxyclineData}
\title{Summary method for oxyclineData}
\usage{
\method{summary}{oxyclineData}(x, ...)
}
\arguments{
\item{x}{Object of class \code{oxyclineData}.}

\item{...}{Extra argumemts.}
}
\description{
Summary method for oxyclineData
}
\examples{
# Set a list of directories
fileMode <- list(fish38_file   = system.file("extdata", "fish38.mat", package = "oXim"),
                 fluid120_file = system.file("extdata", "fluid120.mat", package = "oXim"),
                 blue38_file   = system.file("extdata", "blue38.mat", package = "oXim"))

# Read echograms
echoData <- readEchograms(fileMode = fileMode)

# Calculate oxycline range
echoOutputs <- getOxyrange(fluidMatrix = echoData)

# Get summary
summary(echoOutputs)
}

